# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TAM Rest API",
    "description": "API for the TAM Montpellier public transport system",
    "version": "1.0.0"
  },
  "host": "tam.alexis-mateo.fr",
  "basePath": "/",
  "tags": [
    {
      "name": "Stop",
      "description": "Operations about stops"
    },
    {
      "name": "Line",
      "description": "Operations about lines"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/stop/next/{station}": {
      "get": {
        "tags": [
          "Stop"
        ],
        "summary": "Get next trams at a station. You can filter by line and destination.",
        "description": "",
        "parameters": [
          {
            "name": "station",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "line",
            "in": "query",
            "type": "string"
          },
          {
            "name": "destination",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/stop/schedule/{station}/{line}/{direction}": {
      "get": {
        "tags": [
          "Stop"
        ],
        "summary": "Get the schedule of a station for a line and a direction.",
        "description": "",
        "parameters": [
          {
            "name": "station",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "line",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "direction",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/line/{line}/{direction}": {
      "get": {
        "tags": [
          "Line"
        ],
        "summary": "Get line informations like stops, bounds, etc.",
        "description": "",
        "parameters": [
          {
            "name": "line",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "direction",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/line/": {
      "get": {
        "tags": [
          "Line"
        ],
        "summary": "Get all lines",
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}